json processing fixes
